% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathviewwrap.R
\name{pathviewwrap}
\alias{pathviewwrap}
\title{Title}
\usage{
pathviewwrap(
  fq.dir = "mouse_raw",
  ref.dir = NA,
  phenofile = NA,
  outdir = "results",
  endness = "SE",
  entity = "Mus musculus",
  corenum = 8,
  diff.tool = "DESEQ2",
  compare = "unpaired",
  seq_tech = "Illumina"
)
}
\arguments{
\item{seq_tech}{}
}
\description{
Title
}
