% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sanity_check}
\alias{sanity_check}
\title{Title}
\usage{
sanity_check(
  fq.dir,
  ref.dir,
  phenofile,
  outdir,
  endness,
  entity,
  corenum,
  diff.tool,
  compare
)
}
\arguments{
\item{compare}{}
}
\description{
Title
}
